requirements.txt
django
requests
python-dateutil
pytz
django-admin-tools
